✨ Update Renderer to include structures rendering
Added a vector of Structure references to the Renderer class to enable rendering of structures in the game. Updated the constructor to accept this new parameter and modified the draw method to iterate over the structures and call their render method. This change enhances the rendering capabilities of the Renderer by integrating structure visuals.